Moving more generic functions into the parent class
